requirements.txt
# For loading environment variables
python-dotenv
pandas
openpyxl
geopy
requests
